Check-in before trying to fix DataReader
