+++ ClimateExplorer.Data.AcornSatTransferFunction/HqNewFileParser.cs
